// app/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Difficulty = {
  BEGINNER: 'BEGINNER',
  INTERMEDIATE: 'INTERMEDIATE',
  ADVANCED: 'ADVANCED',
  PROFESSIONAL: 'PROFESSIONAL'
} as const

export type Difficulty = (typeof Difficulty)[keyof typeof Difficulty]


export const QuestionType = {
  MULTIPLE_CHOICE: 'MULTIPLE_CHOICE',
  DRAG_AND_DROP: 'DRAG_AND_DROP'
} as const

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType]


export const ProgressStatus = {
  LOCKED: 'LOCKED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED'
} as const

export type ProgressStatus = (typeof ProgressStatus)[keyof typeof ProgressStatus]